Implementa api para criação de posts no fórum.
